Changed everything to be much more prettier and use signals!
Also added a "no connection" screen.  We're now single threaded again,
which is a good thing (in this case) as we no longer have shared-data
issues.  Now we can use those threads for the actual work: the
image processing.

This "no connection" screen is a compiled-in .xpm file.  I quickly
mocked it up in GIMP.  I had to change the type of the array
from static char* to static const char* to get GCC to shut up though,
so I don't know if it will load it again (I hope...).  Maybe GIMP
needs a patch?